[Tizen][M34-Merge] Fix IME return key type when in Form Tag.
When an user focuses on a text box, the "return" key label on VKB is
shown as "Go" not matter if the focused input a form (submission) or a
simple text input field. Although "Go" is accepted for the former case,
in the case of the latter, Tizen wants to label it as "Done".

So, patch introduces a way to change that from the browser side.

Note that the fix could also apply to android, as it shows the same
behavior, so no Tizen build guards are added.

Original gerrit: http://suprem.sec.samsung.net/gerrit/#/c/2603/
Original Author: Karol Furmaniak <k.furmaniak@samsung.com>

Bug: http://107.108.218.239/bugzilla/show_bug.cgi?id=8415
Reviewed by: Antonio Gomes, SeungSeop Park, Viatcheslav Ostapenko, Youngbok Yoon

Change-Id: Idfbdd5e1057a160db1ace348d73c1d552b6991bd
Signed-off-by: Youngsoo Choi <kenshin.choi@samsung.com>